requirements.txt
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
huggingface-hub==0.25.2
idna==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
requests==2.32.3
safetensors==0.4.5
serpapi==0.1.5
setuptools==75.1.0
smmap==5.0.1
streamlit==1.39.0
sympy==1.13.3
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
transformers==4.45.2
urllib3==2.2.3
